Save album in DB

+++ front-end/src/components/organisms/TripOrganisms/SinglePictureUploader.tsx
@@ -1,32 +1,31 @@
-import { Grid, LinearProgress, makeStyles, Theme } from "@material-ui/core";
+import { faTimes } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { Grid, IconButton, LinearProgress, makeStyles, Theme } from "@material-ui/core";
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Colors } from "../../../styles";
-import {
-  storage,
-  firebase,
-  compressImage,
-  selectIdPerson,
-  AlbumPicture,
-  selectAlbumPictures,
-  getISODatetimeWithOffset,
-  setAlbumPictures,
-  selectAutoDeletePicture,
-  store,
-} from "../../../utils";
+import { compressImage, firebase, selectIdPerson, storage } from "../../../utils";
 import { Text } from "../../atoms";
 
 interface SinglePictureUploaderProps {
+  onUpload: (url: string) => void;
   picture: File;
+  images: File[];
+  updateState: (values: File[]) => void;
 }
 
-export function SinglePictureUploader({ picture }: SinglePictureUploaderProps) {
+export function SinglePictureUploader({
+  picture,
+  onUpload,
+  updateState,
+  images,
+}: SinglePictureUploaderProps) {
   const stylesFunction = makeStyles((theme: Theme) => ({
     colorPrimary: {
       backgroundColor: "#3bab4252",
     },
     progressBar: {
-      width: "85%",
+      width: "75%",
       borderRadius: 10,
       "& .MuiLinearProgress-barColorPrimary": {
         backgroundColor: Colors.GREEN,
@@ -41,29 +40,25 @@ export function SinglePictureUploader({ picture }: SinglePictureUploaderProps) {
 
   const dispatch = useDispatch();
 
+  const [isDeleted, setIsDeleted] = useState(false);
   const [progress, setProgress] = useState(0);
   const idPerson: number = useSelector(selectIdPerson);
-  const autoDeletePicture: boolean = useSelector(selectAutoDeletePicture);
   const userTripRef: firebase.storage.Reference = storage
     .ref()
     .child(`images/trips/${idPerson}`);
-  const imageRef: firebase.storage.Reference = userTripRef.child(picture.name);
+
+  let imageRef: firebase.storage.Reference = userTripRef.child(picture.name);
 
   useEffect(() => {
     uploadPicture();
   }, []);
-
-  useEffect(() => {
-    if (autoDeletePicture) {
-      deletePicture();
-    }
-  }, [autoDeletePicture]);
 
   async function uploadPicture() {
     let fileToSave: File = picture;
 
     if (picture.size > 250000) {
       fileToSave = await compressImage(picture);
+      imageRef = userTripRef.child(fileToSave.name);
     }
 
     let uploadTask = imageRef.put(fileToSave);
@@ -73,48 +68,67 @@ export function SinglePictureUploader({ picture }: SinglePictureUploaderProps) {
         setProgress((snapshot.bytesTransferred / snapshot.totalBytes) * 100);
       },
       (error) => {},
-      async () => {
-        let url: string = await uploadTask.snapshot.ref.getDownloadURL();
-
-        let newAlbumPicture: AlbumPicture = {
-          idPicture: null,
-          pictureUrl: url,
-          date: getISODatetimeWithOffset(new Date(picture.lastModified)),
-        };
-
-        const albumPictures: AlbumPicture[] = store.getState().tripSlice.albumPictures;
-        let updatedAlbumPictures: AlbumPicture[] = [...albumPictures, newAlbumPicture];
-        store.dispatch(setAlbumPictures(updatedAlbumPictures));
+      () => {
+        onUploadSuccess(uploadTask);
       }
     );
   }
 
-  function deletePicture() {
-    imageRef
-      .delete()
-      .then(() => {
-        console.log(`Picture ${picture.name} deleted.`);
-      })
-      .catch((error) => {
-        console.log(`Couldn't delete ${picture.name} | Error: ${error}`);
-      });
+  async function onUploadSuccess(uploadTask: firebase.storage.UploadTask) {
+    let url: string = await uploadTask.snapshot.ref.getDownloadURL();
+    onUpload(url);
   }
+
+  async function deletePicture() {
+    if (!isDeleted) {
+      let fileToDelete: File = picture;
+
+      if (picture.size > 250000) {
+        fileToDelete = await compressImage(picture);
+        imageRef = userTripRef.child(fileToDelete.name);
+      }
+      console.log(imageRef.fullPath);
+      imageRef
+        .delete()
+        .then(() => {
+          setIsDeleted(true);
+
+          let newImages: File[] = images.filter((image) => image.name !== picture.name);
+          updateState(newImages);
+
+          console.log(`Picture ${picture.name} deleted.`);
+        })
+        .catch((error) => {
+          console.log(`Couldn't delete ${picture.name} | Error: ${error}`);
+        });
+    }
+  }
+
+  const formatAsInt = Intl.NumberFormat("en-US", {
+    style: "decimal",
+    maximumFractionDigits: 0,
+  }).format;
 
   return (
     <div>
       <Text style={{ fontSize: "14px", marginBottom: 0 }}>{picture.name}</Text>
 
       <Grid container alignItems="center">
+        <IconButton disabled={progress !== 100} onClick={() => deletePicture()}>
+          <FontAwesomeIcon icon={faTimes} size="xs" color={Colors.BLUE} />
+        </IconButton>
+
         <LinearProgress
           className={style.progressBar}
           classes={{ colorPrimary: style.colorPrimary }}
           value={progress}
           variant="determinate"
         />
+
         <Text
           className={style.percentText}
           style={{ marginBottom: 0, marginLeft: 5 }}
-        >{`${progress} %`}</Text>
+        >{`${formatAsInt(progress)} %`}</Text>
       </Grid>
     </div>
   );
